\name{collate.tags}
\alias{collate.tags}
\title{Builds a list of post tags and links to their associated posts
Extracts tags from the top of all .md files in a list of posts}
\usage{
  collate.tags(posts)
}
\arguments{
  \item{posts}{a character vector of names of posts}
}
\value{
  a list of tag names, each with associated vectors of
  titles and urls for associated posts
}
\description{
  Builds a list of post tags and links to their associated
  posts Extracts tags from the top of all .md files in a
  list of posts
}

