\docType{data}
\name{container_tags}
\alias{container_tags}
\title{A vector of elements that need an explicit endin tag}
\format{chr [1:41] "a" "article" "aside" "b" "body" "canvas" "dd" "div" "dl" "dt" "em" "fieldset" "footer" "form" "h1" ...}
\usage{
  container_tags
}
\description{
  A vector of elements that need an explicit endin tag
}
\keyword{datasets}

