\name{doc.index.list}
\alias{doc.index.list}
\title{Produces an unordered list of links to rendered documentation files}
\usage{
  doc.index.list(doc.pages, out.path)
}
\arguments{
  \item{doc.pages}{a list of Samatha.Page objects, as
  produced by render.all.docs()}

  \item{out.path}{path to documents relative to the Samatha
  site}
}
\value{
  character string representation of an unorderd list of
  links to rendered documentation files
}
\description{
  Produces an unordered list of links to rendered
  documentation files
}

