\name{doc.to.html}
\alias{doc.to.html}
\title{Converts an Rd file to an html fragment}
\usage{
  doc.to.html(doc)
}
\arguments{
  \item{doc}{a path to an Rd file}
}
\value{
  list containing a named character vector of the
  corresponding HTML fragment
}
\description{
  Converts an Rd file to an html fragment
}

