\docType{data}
\name{doctype}
\alias{doctype}
\title{Samatha pages
Functions for wrapping html in doctype boilerplate}
\format{List of 4
 $ html4             :List of 2
 $ xhtml.strict      :List of 2
 $ xhtml.transitional:List of 2
 $ html5             :List of 2}
\usage{
  doctypes
}
\description{
  List of document type declarations for html and xml
}
\details{
  html doctype header represntation
}
\keyword{datasets}

