\name{file.states}
\alias{file.states}
\title{Gets modification times for a vector of files}
\usage{
  file.states(files)
}
\arguments{
  \item{files}{character vector of file paths}
}
\value{
  a named vector of modification times with file paths as
  namess
}
\description{
  Gets modification times for a vector of files
}

