\name{html.postlist}
\alias{html.postlist}
\title{builds a list of links to all blog posts. List is in decreasing order of post date}
\usage{
  html.postlist(site, header.type)
}
\arguments{
  \item{site}{absolute path to the Samatha site}
}
\value{
  character vector representing an unordered html list of
  links to blog posts in descending date order
}
\description{
  builds a list of links to all blog posts. List is in
  decreasing order of post date
}

