\name{html.taglist}
\alias{html.taglist}
\title{returns a character string of an html formatted list of tags,
with numbers of posts associated with each and links to the tag pages
This is a useful addition to an index page}
\usage{
  html.taglist(site)
}
\description{
  returns a character string of an html formatted list of
  tags, with numbers of posts associated with each and
  links to the tag pages This is a useful addition to an
  index page
}

