\name{html_body}
\alias{html_body}
\title{Wraps input in body tags}
\usage{
  html_body(..., opts = list())
}
\arguments{
  \item{\dots}{optional tag body}

  \item{opts}{optional html attributes}
}
\description{
  Build an html body
}
\examples{
html_body("Hello world!")
}
\seealso{
  { m for generating generic html html_head for wrapping
  input in head tags }
}

