\name{html_head}
\alias{html_head}
\title{Wraps input in head tags}
\usage{
  html_head(title, ..., opts = list())
}
\arguments{
  \item{title}{character string}

  \item{\dots}{optional tag body}

  \item{opts}{optional html attributes}
}
\description{
  Build an html header
}
\examples{
html_head("My first page")
}
\seealso{
  { m for generating generic html html_body for wrapping
  input in body tags }
}

