\name{import.tagfile}
\alias{import.tagfile}
\title{Reads in a  JSON tagfile}
\usage{
  import.tagfile(tagfile)
}
\arguments{
  \item{tagfile}{path to JSON tag list file}
}
\value{
  list of tags and associated vectors of posts and links,
  or NULL if file doesn't exist
}
\description{
  Reads in a JSON tagfile
}

