\name{include.markdown}
\alias{include.markdown}
\title{renders the contents of a markdown file as an html character vector
A wrapper around markdown::markdownToHTML}
\usage{
  include.markdown(md.file)
}
\arguments{
  \item{md.file}{a post file in markdown format}
}
\value{
  a character vector of length 1 of rendered html
}
\description{
  renders the contents of a markdown file as an html
  character vector A wrapper around
  markdown::markdownToHTML
}
\seealso{
  include.textfile
}

