\name{include.textfile}
\alias{include.textfile}
\title{renders the contents of a markdown file as an html character vector
A wrapper around readChar}
\usage{
  include.textfile(text.file)
}
\description{
  renders the contents of a markdown file as an html
  character vector A wrapper around readChar
}

