\name{itemise.post}
\alias{itemise.post}
\title{wraps the details of a post up as an rss item
tags and title are extracted from the post for the item description}
\usage{
  itemise.post(post, domain)
}
\arguments{
  \item{post}{a .md file of a blog post}

  \item{domain}{the domain name of the site}
}
\value{
  a character sting of length 1 representing a post in rss
  format
}
\description{
  wraps the details of a post up as an rss item tags and
  title are extracted from the post for the item
  description
}

