\name{make.samatha.page}
\alias{make.samatha.page}
\title{Contstuct a page from the individual elements}
\usage{
  make.samatha.page(content, title, layout)
}
\arguments{
  \item{content}{the actual contents of the page}

  \item{title}{the title of the page, used if your template
  uses it}

  \item{layout}{which layout file should be used}
}
\value{
  Object of class Samatha.Page
}
\description{
  Contstuct a page from the individual elements
}

