\name{render.tagfiles}
\alias{render.tagfiles}
\title{Renders new html pages listing the associated posts for each tag
loops over elements of a taglist, building pages for each tag}
\usage{
  render.tagfiles(site, tag.layout)
}
\arguments{
  \item{site}{the absolute path to the Samatha site}

  \item{tag.layout}{name of the layout file to be used to
  render tags}
}
\value{
  NULL writes to files as a side effect
}
\description{
  Renders new html pages listing the associated posts for
  each tag loops over elements of a taglist, building pages
  for each tag
}

