\name{samatha.compare_state}
\alias{samatha.compare_state}
\title{Compare two directory states.}
\usage{
  samatha.compare_state(old, new)
}
\arguments{
  \item{old}{previous state}

  \item{new}{current state}
}
\value{
  list containing number of changes and files which have
  been \code{added}, \code{deleted} and \code{modified}
}
\description{
  Compare two directory states.
}
\keyword{internal}

