\name{samatha}
\alias{samatha}
\title{Samatha: Runs an infinite loop, updating the site as necessary
This is the main command to update your site.  You can leave this runningg while you make edits to source files}
\usage{
  samatha(site)
}
\arguments{
  \item{site}{character absolute path to your Samatha site}
}
\description{
  Samatha: Runs an infinite loop, updating the site as
  necessary This is the main command to update your site.
  You can leave this runningg while you make edits to
  source files
}

