\name{setup_example_site}
\alias{setup_example_site}
\title{Sets up an example site}
\usage{
  setup_example_site(site, resources = "resources")
}
\arguments{
  \item{site}{path to the directory you want to set up}

  \item{resources}{path to the resources directory in the
  Samatha package}
}
\description{
  provides a barebones Samatha site
}

