\name{tag.links}
\alias{tag.links}
\title{create a set of links for tags without counts}
\usage{
  tag.links(taglist)
}
\arguments{
  \item{taglist}{character vector of tags to generate links
  to}
}
\value{
  a set of fully formatted links to insert into a document
}
\description{
  create a set of links for tags without counts
}

