\name{unordered.list}
\alias{unordered.list}
\title{unordered list representaion}
\usage{
  unordered.list(elements, list.opts = list(),
    element.opts = list())
}
\description{
  Wraps a list of strings into an unordered list
}
\examples{
elements = list("apples", "oranges", "bananas")
unordered.list(elements)
}

