\name{write.docs}
\alias{write.docs}
\title{Renders a directory of Rd files to html and writes within a Samatha site directory}
\usage{
  write.docs(site, doc.path, layout, out.path)
}
\arguments{
  \item{site}{character path to the Samatha site}

  \item{doc.path}{path to the directory of Rd files}

  \item{layout}{Samatha layout file to render documentation
  with}

  \item{out.path}{path to documents relative to the Samatha
  site}
}
\description{
  Renders a directory of Rd files to html and writes within
  a Samatha site directory
}

