\name{write.html}
\alias{write.html}
\title{Writes the html content of a Samatha.Page object to
the file specified in the file element.}
\usage{
  write.html(samatha.page)
}
\arguments{
  \item{samatha.page}{an object of class Samatha.Page}
}
\value{
  logical FALSE if the object is not a Samantha.Page
  object, otherwise TRUE
}
\description{
  Writes the html content of a Samatha.Page object to the
  file specified in the file element.
}

