\name{write.tags.to.file}
\alias{write.tags.to.file}
\title{writes a json file of post tags and links to their associated posts
Seperate from collate.tags to maintain functional style}
\usage{
  write.tags.to.file(site)
}
\arguments{
  \item{path}{to the Samatha site}
}
\value{
  NULL writes to file as a side effect
}
\description{
  writes a json file of post tags and links to their
  associated posts Seperate from collate.tags to maintain
  functional style
}
\seealso{
  collate.tags
}

