\name{extract.tags}
\alias{extract.tags}
\title{Extracts the tag names from a markdown post file
tags should be the first line of the file and preceeded with the percent sign
tags are delimited by spaces, commas or semicolons}
\usage{
  extract.tags(md.file)
}
\description{
  Extracts the tag names from a markdown post file tags
  should be the first line of the file and preceeded with
  the percent sign tags are delimited by spaces, commas or
  semicolons
}

