\name{include.markdown}
\alias{include.markdown}
\title{renders the contents of a markdown file as an html character vector
A wrapper around markdown::markdownToHTML}
\usage{
  include.markdown(md.file)

  include.textfile(text.file)
}
\description{
  renders the contents of a markdown file as an html
  character vector A wrapper around
  markdown::markdownToHTML

  renders the contents of a markdown file as an html
  character vector A wrapper around readChar
}

