\name{include.textfile}
\alias{include.textfile}
\title{reads a text file into a character vector
A wrapper around readChar}
\usage{
  include.textfile(text.file)
}
\arguments{
  \item{text.file}{a text file you wish to include in an
  html page, e.g. a javascript function}
}
\value{
  a character vector of length 1
}
\description{
  reads a text file into a character vector A wrapper
  around readChar
}
\seealso{
  include.markdown
}

