\name{javascript.tag}
\alias{javascript.tag}
\title{Functions for creating HTML elements. These are all wrappers around the html function}
\usage{
  javascript.tag(script)
}
\description{
  Wraps the script string in script tags and a CDATA
  section
}
\examples{
javascript.tag("Some javascript")
}

