\name{ordered.list}
\alias{ordered.list}
\title{ordered list representaion}
\usage{
  ordered.list(elements, opts = list(),
    element.opts = list())
}
\description{
  Wraps a list of strings into an ordered list
}
\examples{
elements = list("apples", "oranges", "bananas")
ordered.list(elements)
}

