\name{render.post}
\alias{render.post}
\title{Render a post from an R markdown file}
\usage{
  render.post(site, postname, layout = "default.R",
    fig.path = "img")
}
\description{
  Render an .Rmd file into a page according to its layout
  template post templates are stored in site/template
  Better date functionality
}

