\name{samatha.engine}
\alias{samatha.engine}
\title{Watches the site directory for changes and recompiles html appropriately}
\usage{
  samatha.engine(site, post.layout = "default.R",
    figure.path = "img")
}
\description{
  Watches the site directory for changes and recompiles
  html appropriately
}

